Change the codebase to be more idiomatic and follows conventions
- Remove `Canonical` and `canon` flag

This library is not supposed to be compiled to WASM or be used by any
contracts.
To be fair we probably won't need to have this library as `no_std`
either (we also have only tests for `std`), but I kept as `no_std`
for future reference, since it helped to some some issue related to
compile a static library on `no_std` environment.

- Implement `to_bytes` and `from_bytes` methods from `Serializable` trait
- Remove `serializable_size()` method in favor of `T::SIZE` from `Serializable` trait
- Add `dusk_bytes::Error` conversion to `Error`
- Use `core::fmt::Display` instead of `std::fmt::Display`
- Change `SecretKey::new` to `SecretKey::random` to be consistent
- Remove the patch version from `dusk_bytes` dependency in `Cargo.toml`
- Change integration tests to be outside `src` as they should be
- Add `rust-toolchain` to ensure CI and local dev behave the same
- Remove overrides from the CI (it will use the `rust-toolchain`)
- Change `rustfmt.toml` to have the same settings of our repos

Resolves: #29