Enable uncompressed deserialisation for PublicKey/APK (#32)
There is no need to use a secure deserialisation when bytes are considered trusted and performance is critical. Therefore, it makes sense to add the capabilities to serialise/deserialise the raw uncompressed keys. Furthermore, the same optimisation shall be applied during the aggregation of the keys.

- Add `to_raw_bytes` and `from_slice_unchecked ` to PublicKey
- Add `to_raw_bytes` and `from_slice_unchecked` to APK
- Fix doc on `APK::serialized_size`
- Add deserialisation benches

Add functions to FFI module
- `pk_to_raw`
- `apk_to_raw`
- `generate_keys_with_raw`
- `aggregate_pks_unchecked`

Resolves: #31